board:2093 a point through a transform is Mat4's own verb, and the fourth hand-rolled box is gone
WHAT UNREAL DOES: `FMatrix::TransformPosition` and `TransformVector` are two names, and the
difference between them is the whole point -- a direction that picked up the translation is
a normal that moves when its object does. WHAT RAGE DOES: the same split, `Transform` and
`Transform3x3`. WHERE THEY AGREE THE MATTER IS CLOSED, so this tree gets both names.

`[12 + axis]` in a loop appears EIGHT times across five files, and a grep says "eight
copies of transform-a-point". It is not, and writing that regex would have been CLAUDE.md's
blind-rename trap exactly: SIX read or write the translation column and TWO transform a
point. Four meanings, one spelling. `Translation` · `SetTranslation` · `TransformPoint` ·
`TransformDirection` now carry them, each `constexpr`, and the pair is held by a
`static_assert` that a direction does NOT pick up the translation -- the defect the two
names exist to make impossible to write by accident.

`Live::Volume` was a FOURTH hand-rolled axis-aligned box, after the three board:2038
replaced with `BoxOf`, and it spent a name `include/scenario/Scenario.h` already gives to a
trigger volume -- two things under one word, which is the thing a name is not allowed to
do. It is `Box` now, its `Empty` flag is gone because `Box` answers that from its own
bounds, and the eight-corner transform is `Box::Through`, asserted at compile time to leave
a TRANSLATED box the same size: growth is what ROTATION costs an axis-aligned box, and a
translation that grew one would mean the corners were not all transformed alike.

AND THE HEADER WAS NOT SELF-CONTAINED, which only the move revealed. `include/math/Mat4.h`
names `Vector3` in four signatures and included nothing for it -- it compiled only as long
as every unit happened to pull `Vec3.h` in first. `Box.h` including it reversed that order
in one unit and produced sixteen errors in a header that had built for months. A door
header that depends on its caller's include order is broken on the day it is written; it
just does not say so until something moves.

  make lint  206 -> 204
  the nine pictures are BIT-IDENTICAL across two clean runs: OldTown 9f4aaec3,
  Heidelberg ee3a3dc0, Shibuya 68bc0c2c, CentralPark 0cba33b9, Venice afa89488,
  Jura ead31be5, ZurichPlan 3af04668, Kaiserberg 9d618d82, Koehlbrand f97ff590

THE FAST GATE IS RED AT 104 AND IT WAS RED BEFORE THIS. Every one of them is
`harness/claims/*` -- the repository's own rules, not the engine. Checked rather than
assumed: `EveryTypeNameIsDeclaredOnce` counts 16 duplicated type names against a DECLARED
9, the sixteen are `Attitude Attribute Body Camera Declaration Document Generator Host
Instance Laid Node Request Sampler Scene Surface Value`, none of them appears in this
round's diff, and the `9` has stood unmoved since 9ae6efc8. So seven names were added
without moving the number that exists to make adding one visible. That is board:2094's and
the next round's, and it is written down rather than walked past -- a gate that runs red
across several commits is a gate that guarded nothing during them.

+++ src/render/stages/LightVisibilityStage.cpp
@@ -73,7 +73,7 @@ void LightVisibilityStage::Build(const Vec3 &preView) {
       for (size_t slot = 0; slot < slots && slot < CastsBelow_; ++slot) {
         const double *const model = placed.data() + slot * 16u;
         for (int axis = 0; axis < 3; ++axis) {
-          const double at = model[12 + axis];
+          const double at = model[12 + static_cast<size_t>(axis)];
           least[axis] = at < least[axis] ? at : least[axis];
           most[axis] = at > most[axis] ? at : most[axis];
         }
